fix(qt_ads): pythons stubs match structure of PySide6QtAds
