+++ bec_widgets/widgets/containers/qt_ads/__init__.pyi
@@ -7,60 +7,10 @@ import typing
 from qtpy import QtCore, QtGui, QtWidgets
 from qtpy.QtCore import Signal
 
+from bec_widgets.widgets.containers.qt_ads import ads
+from bec_widgets.widgets.containers.qt_ads.ads import *
+
 # pylint: disable=unused-argument,invalid-name, missing-function-docstring, super-init-not-called
-
-class SideBarLocation(enum.Enum):
-    SideBarTop = ...
-    SideBarLeft = ...
-    SideBarRight = ...
-    SideBarBottom = ...
-    SideBarNone = ...
-
-class eBitwiseOperator(enum.Enum):
-    BitwiseAnd = ...
-    BitwiseOr = ...
-
-class eIcon(enum.Enum):
-    TabCloseIcon = ...
-    AutoHideIcon = ...
-    DockAreaMenuIcon = ...
-    DockAreaUndockIcon = ...
-    DockAreaCloseIcon = ...
-    DockAreaMinimizeIcon = ...
-    IconCount = ...
-
-class eDragState(enum.Enum):
-    DraggingInactive = ...
-    DraggingMousePressed = ...
-    DraggingTab = ...
-    DraggingFloatingWidget = ...
-
-class TitleBarButton(enum.Enum):
-    TitleBarButtonTabsMenu = ...
-    TitleBarButtonUndock = ...
-    TitleBarButtonClose = ...
-    TitleBarButtonAutoHide = ...
-    TitleBarButtonMinimize = ...
-
-class eTabIndex(enum.Enum):
-    TabDefaultInsertIndex = ...
-    TabInvalidIndex = ...
-
-class DockWidgetArea(enum.Enum):
-    NoDockWidgetArea = ...
-    LeftDockWidgetArea = ...
-    RightDockWidgetArea = ...
-    TopDockWidgetArea = ...
-    BottomDockWidgetArea = ...
-    CenterDockWidgetArea = ...
-    LeftAutoHideArea = ...
-    RightAutoHideArea = ...
-    TopAutoHideArea = ...
-    BottomAutoHideArea = ...
-    InvalidDockWidgetArea = ...
-    OuterDockAreas = ...
-    AutoHideDockAreas = ...
-    AllDockAreas = ...
 
 class CAutoHideDockContainer(QtWidgets.QFrame):
     def __init__(
